some refactoring
Most important one is probably the addition of the SimpleQueryTask.
This task is a JavaFX async task which is capable of running a plain
query. Once results are in, they are displayed.

Refactored a bit of the ResultTable as well. This is not reflected
to the DrillDownTable yet, but that's a TODO. In any case, much can
be simplified by using SimpleObjectProperty as a TableColumns cell
factory.

Also dabbled a bit with JavaFX CSS styling to put monospaced font
in a textarea for errors/exceptions.

Added the Column class which contains column information. It's not
really necessary right now, but we may need it later...