inform user when connection was closed
For instance, when the connection is closed by the peer or a network
change happened, a popup is emitted informing the user that the
connection has died. The tab gets a different icon and the content is
set to disabled. This has to be finetuned but its better than nothing.