read-only option for alias configuration
This allows us to create a connection which is read-only, forcibly.
When not read-only, we open up the possibility to execute insert, update
and delete statements. This does not yet work with the default "SQL" tab
and the drill-down though, since the QueryTask used expects a resultset.

It can work with custom scripts though, and quite possibly in the future
with a custom tab for insert/update/delete statements, but we'll see.